count subsequence
This is an another modification of the recursion subsequences problem to count the subsequences in any array. In this we only remove the data-structure from every function and even from the main function. We do a little change just REMOVE the PRINTING of the data-structure , return TRUE as 1 and False as 0 , STORED the recursive function in the variable name LEFT and RIGHT and at the end ADD BOTH the variable , So that we can count the SUBSEQUENCE.